# agricultural/api-tests/apifox-openapi-import.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "助农电商平台 API",
    "description": "用户端 / 商家端 / 管理员端 RESTful 接口 - 助农电商平台",
    "contact": {
      "name": "助农平台技术支持",
      "email": "support@agricultural.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
  "paths": {
...
    "/admin/merchants": {
      "get": {
        "tags": ["管理员端-商家管理"],
        "summary": "商家列表",
        "operationId": "adminMerchantList",
        "parameters": [
          {"name": "auditStatus", "in": "query", "schema": {"type": "integer"}},
          {"name": "status", "in": "query", "schema": {"type": "integer"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultMerchantInfo"}}}}
        }
      }
    },
    "/admin/merchants/{id}/status": {
      "put": {
        "tags": ["管理员端-商家管理"],
        "summary": "更新商家状态",
        "operationId": "updateMerchantStatus",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {"type": "integer", "description": "0-未开业 1-正常营业 2-禁用"}
                },
                "required": ["status"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
    "/admin/orders": {
      "get": {
        "tags": ["管理员端-订单管理"],
        "summary": "订单列表",
        "operationId": "adminOrderList",
        "parameters": [
          {"name": "status", "in": "query", "schema": {"type": "integer"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultOrderMain"}}}}
        }
      }
    },
    "/admin/orders/{id}": {
      "get": {
        "tags": ["管理员端-订单管理"],
        "summary": "订单详情",
        "operationId": "adminOrderDetail",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultOrderMain"}}}}
        }
      }
    },
    "/admin/orders/{id}/cancel": {
      "put": {
        "tags": ["管理员端-订单管理"],
        "summary": "取消订单",
        "operationId": "adminCancelOrder",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
    "/admin/news": {
      "get": {
        "tags": ["管理员端-新闻管理"],
        "summary": "新闻列表",
        "operationId": "adminNewsList",
        "parameters": [
          {"name": "categoryId", "in": "query", "schema": {"type": "integer"}},
          {"name": "auditStatus", "in": "query", "schema": {"type": "integer"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultNews"}}}}
        }
      },
      "post": {
        "tags": ["管理员端-新闻管理"],
        "summary": "新增新闻",
        "operationId": "addNews",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "categoryId": {"type": "integer"},
                  "content": {"type": "string"},
                  "coverImg": {"type": "string"},
                  "auditStatus": {"type": "integer"}
                },
                "required": ["title", "categoryId", "content"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultNews"}}}}
        }
      }
    },
    "/admin/news/{id}": {
      "get": {
        "tags": ["管理员端-新闻管理"],
        "summary": "新闻详情",
        "operationId": "adminNewsDetail",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultNews"}}}}
        }
      },
      "put": {
        "tags": ["管理员端-新闻管理"],
        "summary": "修改新闻",
        "operationId": "updateNews",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "categoryId": {"type": "integer"},
                  "content": {"type": "string"},
                  "coverImg": {"type": "string"},
                  "auditStatus": {"type": "integer"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      },
      "delete": {
        "tags": ["管理员端-新闻管理"],
        "summary": "删除新闻",
        "operationId": "deleteNews",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/admin/news/categories": {
      "get": {
        "tags": ["管理员端-新闻管理"],
        "summary": "分类列表",
        "operationId": "newsCategoryList",
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/NewsCategory"}}}}}
        }
      },
      "post": {
        "tags": ["管理员端-新闻管理"],
        "summary": "新增分类",
        "operationId": "addNewsCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryName": {"type": "string"},
                  "categoryCode": {"type": "string"}
                },
                "required": ["categoryName", "categoryCode"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultNewsCategory"}}}}
        }
      }
    },
    "/admin/news/categories/{id}": {
      "put": {
        "tags": ["管理员端-新闻管理"],
        "summary": "修改分类",
        "operationId": "updateNewsCategory",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "categoryName": {"type": "string"},
                  "categoryCode": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      },
      "delete": {
        "tags": ["管理员端-新闻管理"],
        "summary": "删除分类",
        "operationId": "deleteNewsCategory",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/admin/after-sale": {
      "get": {
        "tags": ["管理员端-售后管理"],
        "summary": "售后列表",
        "operationId": "adminAfterSaleList",
        "parameters": [
          {"name": "afterSaleStatus", "in": "query", "schema": {"type": "integer"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/PageResultAfterSale"}}}}
        }
      }
    },
    "/admin/after-sale/{id}/handle": {
      "put": {
        "tags": ["管理员端-售后管理"],
        "summary": "处理售后",
        "operationId": "handleAdminAfterSale",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "format": "int64"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "agree": {"type": "boolean"},
                  "handleResult": {"type": "string"}
                },
                "required": ["agree"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
    "/admin/after-sale/{no}/messages": {
      "get": {
        "tags": ["管理员端-售后沟通"],
        "summary": "沟通列表",
        "operationId": "adminAfterSaleMessages",
        "parameters": [
          {"name": "no", "in": "path", "required": true, "schema": {"type": "string"}},
          {"name": "pageNum", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "pageSize", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/AfterSaleCommunication"}}}}}
        }
      },
      "post": {
        "tags": ["管理员端-售后沟通"],
        "summary": "发送消息",
        "operationId": "sendAdminAfterSaleMessage",
        "parameters": [
          {"name": "no", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {"type": "string"}
                },
                "required": ["content"]
              }
            }
          }
        },
        "responses": {
          "200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ResultVoid"}}}}
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"type": "object"}
        }
      },
...
      "ResultNews": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"$ref": "#/components/schemas/News"}
        }
      },
      "News": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "title": {"type": "string"},
          "content": {"type": "string"},
          "coverImg": {"type": "string"},
          "viewCount": {"type": "integer"},
          "categoryId": {"type": "integer", "format": "int64"},
          "categoryName": {"type": "string"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
      "PageResultNews": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/News"}}
        }
      },
      "NewsCategory": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "categoryName": {"type": "string"},
          "categoryCode": {"type": "string"}
        }
      },
      "ResultNewsCategory": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"$ref": "#/components/schemas/NewsCategory"}
        }
      },
...
      "AfterSaleCommunication": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "afterSaleNo": {"type": "string"},
          "senderType": {"type": "integer", "description": "1-用户 2-商家 3-管理员"},
          "senderId": {"type": "integer", "format": "int64"},
          "content": {"type": "string"},
          "isRead": {"type": "integer"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
...
      "AfterSale": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "afterSaleNo": {"type": "string"},
          "orderNo": {"type": "string"},
          "userId": {"type": "integer", "format": "int64"},
          "merchantId": {"type": "integer", "format": "int64"},
          "afterSaleType": {"type": "integer"},
          "applyReason": {"type": "string"},
          "afterSaleStatus": {"type": "integer"},
          "handleResult": {"type": "string"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
      "PageResultAfterSale": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/AfterSale"}}
        }
      },
...
      "MerchantInfo": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "phone": {"type": "string"},
          "merchantName": {"type": "string"},
          "contactPerson": {"type": "string"},
          "contactPhone": {"type": "string"},
          "auditStatus": {"type": "integer"},
          "rejectReason": {"type": "string"},
          "status": {"type": "integer"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
      "PageResultMerchantInfo": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/MerchantInfo"}}
        }
      },
      "OrderMain": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "format": "int64"},
          "orderNo": {"type": "string"},
          "userId": {"type": "integer", "format": "int64"},
          "totalAmount": {"type": "number"},
          "orderStatus": {"type": "integer"},
          "receiver": {"type": "string"},
          "receiverPhone": {"type": "string"},
          "receiverAddress": {"type": "string"},
          "merchantId": {"type": "integer", "format": "int64"},
          "createTime": {"type": "string", "format": "date-time"}
        }
      },
      "ResultOrderMain": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "msg": {"type": "string"},
          "data": {"$ref": "#/components/schemas/OrderMain"}
        }
      },
      "PageResultOrderMain": {
        "type": "object",
        "properties": {
          "total": {"type": "integer", "format": "int64"},
          "list": {"type": "array", "items": {"$ref": "#/components/schemas/OrderMain"}}
        }
      },
...
    }
  }
}